requirements.txt
amadeus>=8.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tkcalendar>=1.6.1
